разносим по пакетам функции
