+++ internal/server/handlers/go.mod
@@ -1,0 +1,3 @@
+module handlers
+
+go 1.20
