+++ internal/server/storage/go.mod
@@ -1,0 +1,3 @@
+module storage
+
+go 1.20
